Create Task 2 - Unemployment Analysis with Python
This commit adds the project files for the Unemployment Analysis using Python. It includes the Jupyter notebook that contains the data analysis pipeline, from data cleaning to visualization, and the CSV file with unemployment data. This project aims to uncover trends and insights into unemployment rates over recent years, with a focus on the impact of the COVID-19 pandemic.